[chore] update DTO float --> double

+++ WaplMath/src/main/java/com/tmax/WaplMath/AnalysisReport/dto/SkillStatDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class SkillStatDTO {
-    private float user;
-    private float average;
-    private float top10Tier;
+    private double user;
+    private double average;
+    private double top10Tier;
 }
